dagster-pipeline/pyproject.toml
[project]
name = "dagster_pipeline"
requires-python = ">=3.10,<3.14"
version = "0.1.0"
dependencies = [
    "boto3>=1.40.19",
    "dagster==1.11.7",
    "dagster-postgres==0.27.7",
    "mysql-connector-python==9.5.0",
    "pandas>=2.3.2",
    "polars==1.35.2",
    "pyarrow>=18.0.0",
    "python-dateutil>=2.8.0",
    "connectorx>=0.3.0",
    "requests>=2.31.0",
    # VTL Pipeline dependencies
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-bigquery-storage>=2.35.0",
    "db-dtypes>=1.3.0",
    "psycopg2-binary>=2.9.0",
    "phonenumbers>=8.13.0",
    # Conversation Analysis dependencies
    "openai>=1.0.0",
    "dspy-ai>=2.5.0",
    "langsmith>=0.1.0",
    # ML Pipeline dependencies (effective supply)
    "catboost>=1.2.8,<1.2.9",
    "scikit-learn>=1.4.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "dagster_pipeline"
registry_modules = [
    "dagster_pipeline.components.*",
]


